reajuste de funcionalidade

+++ src/main/java/Universidade/Service/AlunoService.java
@@ -10,8 +10,6 @@ import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.time.Instant;
-import java.util.Date;
 import java.util.Optional;
 
 @Service
@@ -50,7 +48,7 @@ public class AlunoService {
         return alunoRegistro;
     }
     @Transactional
-    public AlunoRegistro update(AlunoRegistro alunoRegistro,Long id) {
+    public AlunoRegistro update(AlunoRegistro alunoRegistro,Integer id) {
         Alunos a = new Alunos();
         try {
             a = alunoRepository.getReferenceById(id);
@@ -72,7 +70,7 @@ public class AlunoService {
         }
 
         return alunoRegistro;
-    }public void deletarAluno(long id) {
+    }public void deletarAluno(Integer id) {
         Optional<Alunos> aluno = alunoRepository.findById(id);
 
         if (aluno != null) {
